requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
forex-python==1.8
idna==3.4
Markdown==3.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
simplejson==3.19.2
six==1.16.0
urllib3==2.0.7
